Fix build issues for ESP8266Audio and Main component on IDF 5.x
- Updated `flake.nix` to patch ESP8266Audio 1.9.9 and suppress compiler errors (-Wno-error=stringop-overflow, -Wno-error=narrowing).
- Added `app_update` dependency to `main/CMakeLists.txt` for OTA support.
- Fixed `config.h` missing issue by copying `config.example.h` in `flake.nix` build phase if needed.
- Fixed `printf` format specifier mismatch in `MotorTask.cpp`.

+++ main/src/MotorTask.cpp
@@ -310,8 +310,8 @@ String MotorTask::getTestJSON() const {
     char buf[128];
     uint16_t pwm = (uint16_t)(fabs(_testData[i].duty) * 1023.0f);
     snprintf(buf, sizeof(buf),
-             "{\"t\":%u,\"tgt\":%u,\"pwm\":%u,\"cur\":%.3f,\"spd\":%.1f}",
-             _testData[i].t, _testData[i].target, pwm, _testData[i].current,
+             "{\"t\":%lu,\"tgt\":%u,\"pwm\":%u,\"cur\":%.3f,\"spd\":%.1f}",
+             (unsigned long)_testData[i].t, _testData[i].target, pwm, _testData[i].current,
              _testData[i].rpm);
     out += buf;
     if (i < _testDataIdx - 1)
